Move encoding util from action-queue up to top-level
