+++ src/actions/NotificationActions.ts
@@ -9,11 +9,11 @@ import ENV from '../../env.json'
 import { asDevicePayload, DeviceUpdatePayload, NewPushEvent } from '../controllers/action-queue/types/pushApiTypes'
 import { asPriceChangeTrigger } from '../controllers/action-queue/types/pushCleaners'
 import { PriceChangeTrigger } from '../controllers/action-queue/types/pushTypes'
-import { base58 } from '../controllers/action-queue/util/encoding'
 import s from '../locales/strings'
 import { notif1 } from '../modules/notifServer'
 import { getActiveWalletCurrencyInfos } from '../selectors/WalletSelectors'
 import { ThunkAction } from '../types/reduxTypes'
+import { base58 } from '../util/encoding'
 import { fetchPush } from '../util/network'
 import { getDenomFromIsoCode } from '../util/utils'
